refactor app.py (4/7): extract comms helpers to helpers/comms.py
get_active_tags, get_valid_comms, get_after_dark_comms move to
helpers/comms.py. The COMMS_FILE and AFTER_DARK_COMMS_FILE path
constants are redefined locally in helpers/comms.py to match what
app.py used (one was a literal 'static/comms.txt' inside the function,
the other a module constant — both now in the helper).

get_active_tags is internal to comms (only callers are the two reader
functions in this file), so it lives here cleanly.

Smoke-tested: /publish renders 200 (Jinja receives comms_list and
after_dark_comms_list from these helpers).
